static cast double

+++ dorado/read_pipeline/ErrorCorrectionMapperNode.cpp
@@ -274,7 +274,7 @@ ErrorCorrectionMapperNode::ErrorCorrectionMapperNode(const std::string& index_fi
 stats::NamedStats ErrorCorrectionMapperNode::sample_stats() const {
     stats::NamedStats stats = stats::from_obj(m_work_queue);
     stats["num_reads_aligned"] = m_alignments_processed.load();
-    stats["num_reads_to_infer"] = m_reads_to_infer.load();
+    stats["num_reads_to_infer"] = static_cast<double>(m_reads_to_infer.load());
     stats["index_seqs"] = m_index_seqs;
     stats["current_idx"] = m_current_index;
     return stats;
